refactor some functions, monkeypatch max_block_size
