+++ examples/src/main/python/pi.py
@@ -18,11 +18,10 @@
 from __future__ import print_function
 
 import sys
-from random import random
 from operator import add
+from random import random
 
 from pyspark.sql import SparkSession
-
 
 if __name__ == "__main__":
     """
@@ -39,7 +38,7 @@ if __name__ == "__main__":
     def f(_):
         x = random() * 2 - 1
         y = random() * 2 - 1
-        return 1 if x ** 2 + y ** 2 <= 1 else 0
+        return 1 if x**2 + y**2 <= 1 else 0
 
     count = spark.sparkContext.parallelize(range(1, n + 1), partitions).map(f).reduce(add)
     print("Pi is roughly %f" % (4.0 * count / n))
